cleaning up app with modules
